提交ftp收集最后目录修改时间问题

+++ plugins/org.radar.datacollection/curlftp.h
@@ -141,7 +141,7 @@ public:
     int getNewFiles(FileInfoList &fileList);
 
     void listFiles(const string &strDir, FileInfoList &fileList);
-    int parseMlsdInfo(const QString &info, FileInfoList &fileList, QStringList &dirList, int iLatestTime, bool bFtpSupportMSDL);
+    qint64 parseMlsdInfo(const QString &info, FileInfoList &fileList, QStringList &dirList, int iLatestTime, bool bFtpSupportMSDL);
 
     bool connectToHost(const char *url, const char *user_pwd);
